Deploy using nginx buildpack
This allows us to define the entire nginx config, including using maps to selectively enable basic authentication, which isn't something we can do using the Staticfile buildpack.

This is closely based on the configuration we use for the Design System, although we've tweaked the config to suit the Frontend docs. For example, at the minute we do not fingerprint CSS or JavaScript files, so we cannot cache them indefinitely.

Co-authored-by: Hanna Laakso <hanna.laakso@digital.cabinet-office.gov.uk>