Bind logit-ssl-drain service to app
Send logs to the existing logit-ssl-drain service we use for the Design System, by binding the service to the app.

This means that the logs will end up in the same Logit instance as the Design System logs, but we can easily filter the logs and the dashboards by app name.