Update to Bionic
In the previous commit, we updated the Gemfile.lock to use Bundler 2.

By default, Travis uses Ubuntu Xenial 16.04, which comes with Ruby 2.5.3p105 as standard.

As per Travis' documentation:

> 'If you’re using a version of Ruby lower than 2.6 and want to use Bundler 2.x, make sure to upgrade to a newer version of RubyGems. On the default version of RubyGems shipped with older versions of Ruby, the Bundler version in Gemfile.lock must match exactly the version being used, or it will error. It’s fixed in more recent RubyGems releases.' [1]

This is the issue we're now seeing in Travis, as the installed bundler version does not exactly match the version in Gemfile.lock.

Rather than having to manually update RubyGems / bundler in a before_install step, we can instead opt to run under Ubuntu Bionic 18.04, where the standard Ruby version is 2.6.5p114, which shouldn't have this bundler problem.

[1]: https://docs.travis-ci.com/user/languages/ruby/