Remove extraneous 'govuk' in path
Hi team,

I've just installed this now using npm for a bit of bank holiday mucking about, and I noticed that the `node_modules` folder doesn't have a `govuk` subfolder. I'm not sure if I've installed it incorrectly, but in case I've spotted a minor bug I've written this tiny PR to update the docs.