Add a page for CloundFoundry health checks
Create a new HTML page which will be built and deployed to ‘/canary.html’,
containing nothing but the words ‘Tweet tweet’ (following the pattern
set by GOV.UK [1]) and also used by the Design System [2].

Turn authentication off for `/canary.html` so that it can be accessed by
the health checks on the cloud apps domain.

Once this work is merged into master, we can merge in https://github.com/alphagov/govuk-frontend-docs/pull/50
which will add a health check to the manifest based on whether the canary
 page returns 200. See https://github.com/alphagov/govuk-frontend-docs/pull/50
for more details on the health check and why this work had to split
into two PRs.

Further detail:
In a break from the GOV.UK underscores naming convention (eg. `__canary__`),
the page here is deployed to `/canary.html`. This is because the build
 fails with file or directory names starting with underscore. This could
be a Middleman issue as underscores mean something different there. As
this is not a user facing page, just using `/canary.html` feels like a
sufficient solution and not worth digging into further.

[1]: gov.uk/__canary__
[2] design-system.service.gov.uk/__canary__/