Add a health check for deploys to the manifest
https://github.com/alphagov/govuk-frontend-docs/pull/48 added the canary
page to check for.

This PR adds a health check to the deployment to check if `/canary.html`
returns 200 (this change had to be split in two PRs because the manifest
will run the health check before doing the deployment - as our existing site
doesn’t yet have the canary it would fail).

I paired with PaaS tech support on testing what happens with failed deploy
on our repo and we were able to verify that the already deployed site would
 continue to return 200 when an “unhealthy” deploy was attempted (our test
case was deleting ‘deploy/public/canary.html`).

I’m told by the PaaS team that v7 of https://github.com/cloudfoundry/cli
will make these steps in the deployment process much easier to manage as
we’ll be able to do our Travis deploy in one line (instead of the current
three) 🤩