Run tests on Travis
Note that we explicitly exclude the `/vendor` folder from Rubocop
because on Travis gems get installed inside the project itself in
`/vendor/bundle/ruby/[VERSION]/gems/` and so Travis tries
 to run Rubocop on that folder. This results in errors when
Travis finds the linter config for other gems which are using
different versions of Rubocop.

This is not a problem locally because gems get installed globally
outside the project.

Co-authored-by: Oliver Byford  <oliver.byford@digital.cabinet-office.gov.uk>