Deploy using CloudFoundry CLI v7
Install version 7 of the CloudFoundry CLI and use it to deploy the site to PaaS.

The cf push command in v7 has a new flag —strategy which ‘can deploy an app without causing downtime using —strategy rolling’ [1] – which is equivalent to the `v3-zdt-push` command. We also no longer need to manually create the app or apply the manifest, which means we no longer need to extract the app path from the manifest.

These changes mirror the changes made in alphagov/learn-to-code [2] which had the same problem.

[1]: https://docs.cloudfoundry.org/cf-cli/v7.html#:~:text=You%20can%20deploy%20an%20app%20without%20causing%20downtime%20using%20cf%20push%20app_name%20--strategy%20rolling
[2]: https://github.com/alphagov/learn-to-code/pull/40