Add CODEOWNERS file to protect secrets
We need to make sure only staff can access secrets and/or deploy to
production.

This commit adds a [CODEOWNERS file] to ensure that any changes to our
GitHub Actions or Travis CI pipelines are reviewed by a developer, to
avoid the (unlikely) scenario where a PR that changes our CI/CD is
approved by someone who isn't a developer and doesn't spot the
significance of the change.

[CODEOWNERS file]: https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/creating-a-repository-on-github/about-code-owners