Fix deploys, take #2
Commit f731fa1 tried to fix broken deploys by preserving the top level
directory name of the deploy folder. This didn't work, files are
still being archived as `public/whatever` instead of
`deploy/public/whatever`. This commit tries a different tack and just
moves the manifest into the deploy folder, and then ensures that only
the deploy folder is present on the machine that does the deploy. This
means that we don't need to worry about pushing things to PaaS that we
don't want to deploy.