Refactor GitHub Actions workflows
Split the `ci.yaml` workflow into two: `test.yaml` for pull request
checks, and `deploy.yaml` for after merge to main. To reduce code
duplication the deploy workflow calls the test workflow using the beta
reusable workflow feature [[1]].

The advantage of splitting the workflows like this is that we can then
give the entire deploy workflow a concurrency limit, without affecting
test workflows.

This is the equivalent of [commit 6f04855e in the govuk-design-system
repo][2].

[1]: https://docs.github.com/en/actions/learn-github-actions/reusing-workflows
[2]: https://github.com/alphagov/govuk-design-system/pull/1971/commits/6f04855e5089d4adb441563b9fb75abc5e041f49