Add upload-artifact input
Changes the test workflow so that rather than only publishing an
artifact when run on the main branch, it is an input argument of the
workflow call [[1]].

The commit adds an input that specifies whether to upload a zip of the
build folder as an artifact. We usually only want to do that when we're
doing a deploy, because we don't want to run setup and build twice, and
we want the deploy to be a separate job, and you can't share folders
between jobs any other way. Previously we would always upload the
artifact, but that can take some time.

This is the equivalent of [commit fac3329 in the govuk-design-system
repo][2].

[1]: https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#onworkflow_callinputs
[2]: https://github.com/alphagov/govuk-design-system/commit/fac33297b13e9e4cfa23d88b46a639dc0cbc25d3