Add concurrency limit to production deploys
Refactors deploy workflow to use a concurrency limit to avoid trying to
run more than one deploy at a time [[1]].

This is an alternative to the step which cancels any in-progress
deploys, which would have resulted in failed deployment messages.

This should make it easier to spot when a deployment fails for real.

This is the equivalent of [commit 336a93 in the govuk-design-system
repo][2].

[1]: https://github.blog/changelog/2021-04-19-github-actions-limit-workflow-run-or-job-concurrency/
[2]: https://github.com/alphagov/govuk-design-system/pull/1971/commits/336a93c78812a06afe6f55ea5850de9e1f5c75aa