Remove unnecessary guidance for webpack
We added this guidance for govuk-frontend v4.1.0 - when we added the ability to use ES Modules - because our imports were not fully specified (ie: they did not have file extensions). We've [fixed this for v4.1.1](https://github.com/alphagov/govuk-frontend/pull/2658), and this guidance now looks to be unneeded.

Part of #194