Work around table of contents bug
There appears to be a bug in the way that the tech docs generates unique identifies for headings.

The link to the 'Button' component in the JS API Reference from the Sass API reference is incorrect because the Sass API Reference _also_ has a heading called 'Button' and this is incorrectly taken into consideration when building the nav even though we're linking to a different page.

Work around this by adding the word component to every H2 to make the headings unique, until we can fix the bug in the tech docs gem.