Fix sidebar menu render for v4
Allows the v4 index to be listed at the same level as the other v4 pages, instead of having them listed as sub items of the index.

Prevents the rendering of the other v4 pages as sub-items of the v4 index to avoid duplication. This is achieved by overriding
the method rendering the tree of links in the menu so children of one specific page can be excluded from being rendered temporarilly.

Co-authored-by: Brett Kyle <brett.kyle@digital.cabinet-office.gov.uk>