Add in-page navigation to v4 table of content
Adds a missing statement from tech_docs_gem's layout.

Unfortunately, because `content_for` appends content to the slot it's filling,
and the version in Middleman does not seem to support the `flush` option for replacing rather than appending,
we can't do `wrap_layout :layout` and override `:sidebar` so have to re-implement the same line as in the layout from tech_docs_gem.