Result of fiddling with tsconfig to get it to generate a typedoc
We were able to get 116 errors down to only 5 with this setup, however we only managed to succeed using the following command:

```
typedoc --skipErrorChecking
```

This will produce a typedoc. To see methods in the typedoc UI, you need to click on 'settings' on the right and check 'external'.

You can also produce json by adding the flag `--json [filename].json`.